Small fix to allow the branch to compile

+++ core/src/gbts_seeding/gbts_seeding_config.cpp
@@ -81,8 +81,8 @@ bool gbts_seedfinder_config::setLinkingScheme(const std::vector<std::pair<int, s
 
 	//contianers sizes
 	volumeMapSize   = largest_volume_index + 1;
-	nLayers         = layerInfo.isEndcap.size();
-	surfaceMapSize  = surfaceToLayerMap.size();
+	nLayers         = static_cast<unsigned int>(layerInfo.isEndcap.size());
+	surfaceMapSize  = static_cast<unsigned int>(surfaceToLayerMap.size());
 	
 	TRACCC_INFO("volume layer map has " << volumeToLayerMap_unordered.size() << " volumes");
 	TRACCC_INFO("The maxium volume index in the layer map is " << volumeMapSize);
